TestHarnass: more robust

+++ dkimproxy/trunk/tests/TestHarnass.pm
@@ -30,34 +30,52 @@ sub start_servers
 	$self->{proxy_port} ||= 20025;
 	$self->{sink_port} ||= 20026;
 
-	$self->{proxy_pid} = fork;
-	if ($self->{proxy_pid} == 0)
-	{
-		exec("$path/dkimproxy.out",
-		@{$self->{proxy_args}},
-		"127.0.0.1:$self->{proxy_port}",
-		"127.0.0.1:$self->{sink_port}",
-		)
-		or die "Error: cannot spawn dkimproxy.out process: $!\n";
-	}
-	elsif (not defined $self->{proxy_pid})
-	{
-		die "fork: $!\n";
-	}
-
+	# launch the SMTP sink
 	my $sink_fh;
 	$self->{sink_pid} = open($sink_fh, "-|",
 			"./smtp_sink.pl", "--port=20026")
 		or die "Error: cannot spawn smtp_sink.pl process: $!\n";
 	$self->{sink_fh} = $sink_fh;
 
-	# give it enough time to start
-	sleep 1;
+	# wait for it to be ready
 	my $tmp = <$sink_fh>;
 	if (not $tmp)
 	{
-		$self->shutdown_servers;
 		die "Error: smtp_sink.pl failed to start\n";
+	}
+
+	# launch DKIMproxy
+	my $pidfile = "/tmp/pidfile$$";
+	$self->{proxy_pid} = fork;
+	if ($self->{proxy_pid} == 0)
+	{
+		exec("$path/dkimproxy.out",
+		@{$self->{proxy_args}},
+		"--pidfile=$pidfile",
+		"127.0.0.1:$self->{proxy_port}",
+		"127.0.0.1:$self->{sink_port}",
+		)
+		or do {
+			$self->shutdown_servers;
+			die "Error: cannot spawn dkimproxy.out process: $!\n";
+			};
+	}
+	elsif (not defined $self->{proxy_pid})
+	{
+		$self->shutdown_servers;
+		die "fork: $!\n";
+	}
+
+	# wait for DKIMproxy
+	my $elapsed = 0;
+	until (-f $pidfile)
+	{
+		sleep 1;
+		$elapsed++;
+		if ($elapsed >= 6) {
+			$self->shutdown_servers;
+			die "Error: dkimproxy.out (pid $self->{proxy_pid}) failed to create PID file\n";
+		}
 	}
 
 	print "# have pids $self->{sink_pid}, $self->{proxy_pid}\n";
@@ -68,19 +86,25 @@ sub shutdown_servers
 {
 	my $self = shift;
 
-	print STDERR "shutting down proxy...\n";
-	kill "TERM", $self->{proxy_pid};
-	wait;
-	delete $self->{proxy_pid};
+	if ($self->{proxy_pid})
+	{
+		print STDERR "shutting down proxy...\n";
+		kill "TERM", $self->{proxy_pid};
+		wait;
+		delete $self->{proxy_pid};
+	}
 
-	print STDERR "shutting down sink...\n";
-	kill "TERM", $self->{sink_pid};
-	wait;
-	my $sink_fh = $self->{sink_fh};
-	close $sink_fh;
+	if ($self->{sink_pid})
+	{
+		print STDERR "shutting down sink...\n";
+		kill "TERM", $self->{sink_pid};
+		wait;
+		my $sink_fh = $self->{sink_fh};
+		close $sink_fh;
+		delete $self->{sink_fh};
+		delete $self->{sink_pid};
+	}
 
-	delete $self->{sink_fh};
-	delete $self->{sink_pid};
 	return;
 }
